!! NOT WORKING !! - removing CoreData - using Realm for database - looping references wallet -> entries -> wallet -> entries

+++ src/Domain/Enums/WalletEntryType.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-public enum WalletEntryType: Int16 {
+public enum WalletEntryType: Int {
     case outcome = 0
     case income
